Replace `HealthManager` with `ConstrainedFloatDisplay`
